refactor(admin): AdminEvaluationDetails sous 300 lignes — éclatement complet (H3)
Décompo de src/views/admin/AdminEvaluationDetails.vue (730 → 196 l) selon le
patron G1 (AdminUsers) : toute la logique extraite dans un composable, sections
de template découpées en sous-composants présentationnels, vue réduite à un
orchestrateur (header + états chargement/erreur + conteneurs).

Créé :
- src/composables/useAdminEvaluationDetails.js (131 l) : state, lecture route id,
  appel api results-by-class, helpers statut/note/dates, exportCSV, onMounted.
- src/components/admin/EvaluationStatsCards.vue (112 l) : 4 cartes stats globales.
- src/components/admin/EvaluationInfoCard.vue (85 l) : infos évaluation (formatDate
  injecté en prop pour parité exacte).
- src/components/admin/EvaluationResultsTable.vue (226 l) : tableau résultats +
  état vide + bouton export (émet `export`), helpers injectés en props.
- src/components/admin/EvaluationDistributionStats.vue (97 l) : distribution +
  répartition.
- tests/unit/useAdminEvaluationDetails.test.js (96 l) + EvaluationResultsTable.test.js (48 l).

Parité : mêmes appels service (api.get /evaluations/:id/results-by-class), même
route (route.params.id), mêmes classes CSS, mêmes textes/formats. CSS déplacé
VERBATIM avec chaque section. `.empty-state/.empty-icon/.empty-text` déplacés dans
EvaluationResultsTable (leur markup y vit).

Dette pré-existante CONSERVÉE à l'identique :
- Le sélecteur `.loading-state` (groupé avec `.error-state`) reste dans la vue
  alors que le chargement utilise <ContentLoader> : CSS mort sans markup, déjà
  présent avant la décompo, non « corrigé » (conservé dans le bloc style de la vue).

Tests: 10/10 verts. Build: exit 0 (warnings chunk-size pré-existants ignorés).
